feat: 로그인(인증) 기능 구현

+++ src/main/java/com/example/schedulemanagementdevelop/schedule/controller/ScheduleController.java
@@ -2,6 +2,7 @@ package com.example.schedulemanagementdevelop.schedule.controller;
 
 import com.example.schedulemanagementdevelop.schedule.dto.*;
 import com.example.schedulemanagementdevelop.schedule.service.ScheduleService;
+import com.example.schedulemanagementdevelop.user.dto.SessionUser;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -20,39 +21,60 @@ public class ScheduleController {
     // 일정 추가
     @PostMapping()
     public ResponseEntity<CreateScheduleResponse> createSchedule(
+            @SessionAttribute(name = "loginUser", required = false) SessionUser sessionUser,
             @PathVariable Long userId,
             @RequestBody CreateScheduleRequest request) {
+        if(sessionUser == null) {
+            throw new IllegalStateException("로그인이 필요한 기능입니다.");
+        }
         return ResponseEntity.status(HttpStatus.CREATED).body(scheduleService.save(userId, request));
     }
 
     // 일정 단건 조회
     @GetMapping("/{scheduleId}")
     public ResponseEntity<GetOneScheduleResponse> getOneSchedule(
+            @SessionAttribute(name = "loginUser", required = false) SessionUser sessionUser,
             @PathVariable Long userId,
             @PathVariable Long scheduleId) {
+        if(sessionUser == null) {
+            throw new IllegalStateException("로그인이 필요한 기능입니다.");
+        }
         return ResponseEntity.status(HttpStatus.OK).body(scheduleService.getOne(userId, scheduleId));
     }
 
     // 일정 전체 조회
     @GetMapping()
     public ResponseEntity<List<GetAllScheduleResponse>> getAllSchedule(
+            @SessionAttribute(name = "loginUser", required = false) SessionUser sessionUser,
             @PathVariable Long userId,
             @RequestParam(required = false) String userName) {
+        if(sessionUser == null) {
+            throw new IllegalStateException("로그인이 필요한 기능입니다.");
+        }
         return ResponseEntity.status(HttpStatus.OK).body(scheduleService.getAll(userId, userName));
     }
 
     // 일정 수정
     @PatchMapping("/{scheduleId}")
     public ResponseEntity<ModifyScheduleResponse> modifySchedule(
+            @SessionAttribute(name = "loginUser", required = false) SessionUser sessionUser,
             @PathVariable Long scheduleId,
             @Validated @RequestBody CreateScheduleRequest request
     ) {
+        if(sessionUser == null) {
+            throw new IllegalStateException("로그인이 필요한 기능입니다.");
+        }
         return ResponseEntity.status(HttpStatus.OK).body(scheduleService.modify(scheduleId, request));
     }
 
     // 일정 삭제
     @DeleteMapping("/{scheduleId}")
-    public ResponseEntity<Void> deleteSchedule(@PathVariable Long scheduleId) {
+    public ResponseEntity<Void> deleteSchedule(
+            @SessionAttribute(name = "loginUser", required = false) SessionUser sessionUser,
+            @PathVariable Long scheduleId) {
+        if(sessionUser == null) {
+            throw new IllegalStateException("로그인이 필요한 기능입니다.");
+        }
         scheduleService.delete(scheduleId);
         return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
     }
